ci(gpg): Fix copied em dash for CLI option
Oddly, this was found by Vale, a prose linter. I thought that the ``--ext=.pl`` option
meant that Value would only lint comments. It runs out that the
``./styles/RedHat/EmDash.yml`` rule uses ``scope: raw`` which is why it caught the em
dash in code. I'm guessing this rule uses that scope because it also checks for
``&mdash;`` and wants to check HTML "code" not just HTML comments.