build(compose): TODO Current best practice updates
Had to add another exception to the `$ ~/.nvm/nvm-exec npm exec -- write-good` linter
for a word in a URL in a reStructuredText document. That brings us pretty close to
disabling half it's checks. Given that it provides no way to lint only prose and not
code or markup nor does it provide any way to ignore certain words, it may be best to
just stop using it. Or maybe we lint only `*.rst` files and render them to text before
piping them in?