fix(docker): Give up on shell history in ~/.config
I realized that even with the previous changes the `./docker-compose.yml`
configuration would result in errors for end users because they don't use the
`./Makefile`. This has all been too much hassle. Go back to a bind mount that isn't a
shared volume which is much simpler and removes an image layer.

This reverts commit c249ffdfa6999a3fbc8058ecdaf6c58fa351bb29.