build(compose): Cleanup repetition w/ YAML anchors
I wish the canonical docs made it clear when to use with of the various methods for
managing complexity and repetition in the configuration of a compose project. Here's
what I've learned so far:

- YAML `&` anchors can only be used within one file
- The `extends:` option can only reference a service and can only be used in a service.

Altogether, this means that only whole services can be shared between files, no `x-*`
extensions and no YAML anchors. That's unfortunate, it would be really useful to be able
to share `x-*` extensions between files.

Finally, regarding plain YAML aliases vs YAML merging, using an alias only saves the
`<<: ` key prefix and using an alias means the config is broken if someone tries to add
overriding keys under the alias. I also find the `<<: ` prefix more readable and it
more clearly signals to me that some template-ish stuff is going on. So I've decided to
avoid adding another syntax and only use the `<<: ` merge prefix.